fix: rename jwt file
